231114_Vue_with_DRF/실습 과제/back/posts/serializers.py
from rest_framework import serializers
from .models import Category, Post, Comment


class PostListSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['pk', 'title', 'category']


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'