231114_Vue_with_DRF/실습 과제/back/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-seed==0.3.1
djangorestframework==3.14.0
Faker==20.0.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
